Removed redundant class and moved CardDrawView to correct package
